+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/data/response/SettingsResponse.kt
@@ -1,0 +1,7 @@
+package com.inventiv.gastropaysdk.data.response
+
+data class SettingsResponse(
+    val faq: String?,
+    val phoneNumber: String?,
+    val email: String?,
+)