+++ gastropaysdk/src/main/java/com/inventiv/gastropaysdk/ui/settings/SettingsFragment.kt
@@ -14,10 +14,12 @@ import com.inventiv.gastropaysdk.repository.SettingsRepositoryImp
 import com.inventiv.gastropaysdk.shared.GastroPaySdk
 import com.inventiv.gastropaysdk.ui.MainViewModel
 import com.inventiv.gastropaysdk.ui.MainViewModelFactory
+import com.inventiv.gastropaysdk.ui.contactus.ContactUsFragment
 import com.inventiv.gastropaysdk.ui.webview.WebViewFragment
 import com.inventiv.gastropaysdk.utils.blankj.utilcode.util.LogUtils
 import com.inventiv.gastropaysdk.utils.blankj.utilcode.util.StringUtils
 import com.inventiv.gastropaysdk.utils.delegate.viewBinding
+import com.inventiv.gastropaysdk.utils.getSettings
 import com.inventiv.gastropaysdk.utils.handleError
 import com.inventiv.gastropaysdk.view.GastroPaySdkToolbar
 import kotlinx.coroutines.flow.collect
@@ -59,6 +61,19 @@ internal class SettingsFragment : BaseFragment(R.layout.fragment_settings_gastro
         }
         binding.termsButton.setOnClickListener {
             viewModel.getTerms()
+        }
+        binding.contactUsButton.setOnClickListener {
+            sharedViewModel.pushFragment(ContactUsFragment())
+        }
+        binding.faqButton.setOnClickListener{
+            getSettings().faq?.let { url ->
+                sharedViewModel.pushFragment(
+                    WebViewFragment.newInstance(
+                        toolbarTitle = StringUtils.getString(R.string.settings_faq_gastropay_sdk),
+                        url = url
+                    )
+                )
+            }
         }
     }